Clean up VRAM to allow for running on 8xV100
I found that the rank 0 gpu needs to have its cache emptied or it will OOM on more than 4 GPUs in the same instance. This fixes that to a degree - although I'm still working on making sure each of the 8 GPU's is being used effectively - as it stands they can only get to about 10GiB/16GiB vram usage with a batch size of 64 (depth 8, heads 8, attn_types=('sparse'), reversible=False). But this should help at least.